bot_stack/bot/requirements.txt
python-telegram-bot==21.4
requests==2.32.3
fastapi==0.112.1
uvicorn==0.30.6
aiofiles==23.2.1
aiohttp==3.9.0